Move Components to proper directory
