+++ src/app/app-routing.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { PageNotFoundComponent } from './components/page-not-found/page-not-found.component';
-import { SoftwarePlatformViewComponent } from './components/software-platforms/software-platform-view/software-platform-view.component';
-import { CloudServiceViewComponent } from './components/cloud-services/cloud-service-view/cloud-service-view.component';
+import { SoftwarePlatformViewComponent } from './components/execution-environments/software-platforms/software-platform-view/software-platform-view.component';
+import { CloudServiceViewComponent } from './components/execution-environments/cloud-services/cloud-service-view/cloud-service-view.component';
 import { AlgorithmListComponent } from './components/algorithms/algorithm-list/algorithm-list.component';
 import { PublicationListComponent } from './components/publications/publication-list/publication-list.component';
 import { ImplementationViewComponent } from './components/algorithms/implementation-view/implementation-view.component';
