+++ src/app/components/execution-environments/software-platforms/software-platform-view/software-platform-view.component.ts
